Added reset status register for 908 to device definition
I thought I needed this, to check (in the chat loader) if what caused
reset and to vary behaviour accordingly. It doesn't work that well, and
it's ugly code, so I'm sticking with my simple but mildly broken chat
code.

The main problem: on the S08 in particular, touching an undefined
address will cause an illegal address (ILAD) abort (reset). When dumping
memory, this is annoying. Fortunately, I got lucky: the ILAD reset would
end up in the SWI code, which ends by returning the flags register. So
the dump would get _something_.

That's how I'm going to leave it. There is more to the ugliness of the
behaviour, but I don't want to talk about it... Mostly JS16's that
return _multiple_ bytes (and I don't understand why).