Improve robustness of switching between chat and bdm
My resync code - which flushes the tty input buffer and then sends two
"idle" bytes (which, if the BDM code is running, cause it to exit) -
didn't work reliably, because I wasn't waiting for the characters to go
down the line, and possibly tickle a response from the target.

If the BDM _is_ running, resync will cause it to exit, and it will
return to the "SWI" handler in chat, which will send a byte. But this
won't happen immediately. With this change, after sending the idles
bytes, we wait for the tty _output_ buffer to drain, and then continue
on to the chat code, which will resync again the first time it tries to
read the SP or set a memory address; at this point it will successfully
throw out the response byte.