Reverted (mostly) to older RAM-routine for flashing S08 chips
This commit essentially "reverts"
95402c96318635d531cac393a52809af2bcb3aea, but makes some other
improvements as well.

I didn't want to wholesale revert that commit because there were changes
that came along with it that I wanted to keep (I forget now what they
were).

I'm also using simpler and shorter RAM routine to write the flash;
including the byte (row) count, it's only six bytes longer than the
"streamlined" version that I was using to write single bytes, and
scripting via chat ... and which was very slow.

Now flash via both bdm and chat are fast again.

This commit also brings in some niceties wrt flashing via bdm, which,
since we're executing code, requires that the watchdog be turned off -
in addition to trimming the osc and setting fcdiv. All these things are
done "automatically" - and in addition the RAM flashing routine is
copied over - if we are able to successfully connect via bdm and get a
c8 status back.

It's pretty nice.

And treatment of registers - while still using the (somewhat bogus) idea
of a host-side cache of all the registers, now works nicely and is kept
better synchronised - esp when switching between chat (which shows the
registers of the _chat_ target) and bdm (which shows the registers of
the connected bdm target).

There are still things to clean up, but this is the main semantic change
that gets us back to fast flashing.