Using bgnd broke ability to flash bootloader
In order to flash the bootloader, you have to load a RAM-resident
version and execute that, then re-load the code on the host with the
version targeted at flash, and flash it.

Trouble is, now that we're using a RAM routine to write flash - for both
chat and bdm - and using a bgnd instruction at the end - we take an
illegal opcode reset after the code gets executed... which means after
erasing the secure byte, the trims, and the vectors, and then the reset
has no where to go... the chip is dead, and we have to go thru a
complicated ritual to resuscitate it. This is a drag.

We need to have a way to distinguish between flashing via chat (which
will exit via rts) and via bdm (which exits via bgnd). The easiest way
is simply to patch the routine every time we use it. I'm doing this
already, to paste in the flash command and byte count. So now I paste in
an instruction as well. And it gets defined by a deferred command,
t.Return.. which BDM defines to return bgnd, and chat defines to return
rts. Simple!