BDM now smarter about not setting trim if not found in flash
QE parts in particular have a quirk that on normal POR they set the trim
from a secret location (not in the flash array) but if brought up in
forced bdm mode, they leave the trim @ 080h, and there is no way to
access the factory trim.

Rather than blindly using the 0ff "trim" byte at ffaf - which can tweak
the oscillator out of the range of bdm comms - we print a message and
let the user figure out what to do. Both TRIM and FCDIV need to be set
manually before the chip can be flashed.