Added flash protection and vector redirection; fixed flash routine again
Added support for S08 (of the newer variety, anyway - like JS, JM, QE,
and QG; the really old stuff, like GB is a bit different) flash
protection and vector redirection.

We default to protecting the chat loader - the last 512 bytes of flash
(0fe00). And we enable vector redirection, so all but the reset vector
(which stays put) get moved down 512 bytes. SWI vector is now at 0fdfc,
for example.

It's easy to protect more flash:

   <first protected address> protect-flash

This automatically calclulates the vector redirection - but it has to be
executed _before_ any IRQ handlers are compiled (the redirection is
applied at the time that they are compiled).

While doing all this I discovered that my RAM-based flash routine still
isn't quite up to scratch. In particular, I was bitten again by
returning to code in flash before the flash charge pump has turned off.
You really _must_ wait for the FCCF bit in FSTAT (bit 6) before you can
exit RAM. Now the code is more careful about this.

I could have simply given up on using burst program commands, but I
wanted to see what was involved. If I changed to that, the code would
get even simpler; I could elide the loop that waits for FCCF at the end.

One curious feature of the existing code: the last FSTAT value that was
valid is returned shifted left by one: the high bit is in C, and bit 0
is 0. To "reconstitute" it (on the host) we concatenate CC with A (and
get lucky - C is the lsb of the flags register), shift the whole thing
right, and then mask off the low byte. We use this value to test for
flash access or (protection) violation errors.