Added TPM and SCI synonyms to several chip definitions
The chat code uses an SCI, and the BDM code uses both an SCI and a TPM.
For chips that have only one of each, we use that one. For chips that
have more than one, we have to create synonyms.

I simply added these at the end of the chip defs.

In the process I moved the definitions of "reg" and "vector" to the main
load file and fixed a subtle bug: having \ -- at the end of both reg
and vector requires that there be text to consume after the name; if
there isn't, -- actually eats the entire _next_ line up to the newline.
Why? Because if -- is at the end of a line, the parser has consumed the
newline in parsing it; when it executes, it grabs the input stream up to
the _next_ newline, which means it's going to eat the entire following
line!

I got bit by this because my synonym definitions didn't have any
comments, and only every other one of them was getting compiled. I was
puzzled by this, until I realised this quirk of --.

Now I'm using | to consume the bit field comments after reg name
definitions. I should probably use a longer and less useful single char
token...