Inching toward S08 metacompiler!
Removed all mentions of inside, outside, \o and \i.

Changed code to compile into target, and so changed all my uses of code
in the chat and bdm code into label - since I'm using the asm as an
assembler, and not compiling forth code.

Changed a few naming things in memory.mu4 - mostly got rid of the |
words: |here etc.