Interactive test of S08 forth words!
I added a few more words to the S08 kernel, and wrote simple interaction
code, which pushes the host stack to the target, executes a single word,
and pops the target stack back across.

I've tested all the words in my (incipient) S08 kernel, and they all
work.

Oddly, I discovered something that I didn't know I didn't know. ;-)

In the process of implementing u< and < I discovered that my hunch -
that the conditions codes after the last byte of a multi-byte subtract
can't represent the state of the _whole_ result - was _correct_.

I guess I had never thought about it - and I've never actually
implemented "multiprecision" (ie, doing several chained
subtract/subtract-with-carry instructions) u< and <. fig "cheats" and
implements a funny "modulo" <.  Their definition is

  : <   - 0< ;

which is _neither_ u< nor <. And both the V25 and 6809 - the two
architectures for which I've written 16-bit Forth kernels - have
16-bit add and subtract instructions, so I never had to worry about
doing it byte-wise. Lastly, the ARM - which, while its idea of carry is
backwards from most architectures, has 32-bit subtracts, so this point
was moot there as well.

Why was any of this even an issue? It turns out that, because of the way
I represent literals in the HC08 forth, I couldn't do the subtract the
right way for u< and <. We want to calculate second minus top; but with
my representation I could only calculate top minus second - so instead
of using u< (aka BLO (branch lower) aka branch if carry set) I used u>
(aka BHI (branch if higher) aka branch if carry and zero both unset).

The trouble? Z. Its value after the most-significant subtract only
represents _that_ byte, not the whole value! We propagate carry but not
zero ... and this was the intuition that was bothering me. It turned to
be totally correct: the code using BHI didn't work. At all.

< is similar. We want to use < (aka BLT aka branch if N xor V == 1) but
because of the issue with the order of operands I used > (aka BGT aka
branch if Z or (N xor V) == 0) - and, because of the presence of the Z
flag in that test, that didn't work either.

By treating u< and < differently from the other 2ops (+ and xor or) for
which I handle literals specially) I was able to calculate the subtract
in the correct order to be able to use conditional branches that don't
test Z (ie BLO and BLT). Doing this makes the tests work perfectly.

The moral? If you're doing multi-word (where "word" is the native data
chunk size of your machine) compares to test for less than, you have do
the subtract in the right order so you can use conditional branch
instructions that don't test the Z flag, which won't represent the whole
value. This applies to more than just 8-bit micros: if I'm doing a
32-bit u< on a 16-bit device, or 64-bit u< on a 32-bit device, I have to
be careful about this.

I never knew. Until now.