interact supports both one- and two-byte scratch hc08 kernels
I'm not sure I'm going to be able to use one-byte scratch versions of
the hc08 kernel. Even though it's an elegant formulation, because 16-bit
values are little-endian on the stack, I can't use ldhx or sthx ...
which I need for @ ! and friends.

But I wanted to make sure both versions of the kernel work, and before
this commit, the interact code was "wired" for the two-byte-scratch
kernel. Now it's smarter about working with both.