Added .lex. to meta-compiler chains - comments and cond interpretation
There is a handful of words that are always useful, esp in interpreter
modes that don't search .forth. and so miss out on the utility words in
there.

Having the comment parsing words  (  and  --  is always handy. And being
able to .ifdef something is as well.

For any mode that wants them, these are all in .lex. now.