Optionally show bytes being comma'ed into target image
For months (years?) now I've had the HC08 image code spew out the bytes
it compiles into the image. This is somehow comforting, and it's easy to
see if something has gone wrong (assuming you can disassemble hc08
machine code in your head - which I was getting pretty good at before I
got my disassembler up to scratch).

But it's also noisy. I've added a define - show-comma - that is tested.
If undefined, nothing is printed - so the default is to be quiet.