Added shift-and-subtract divide routine to HC08 kernel
Unsigned 32/16->16/16 divide.

And I learned another quirk of computer arithmetic - you have to keep
track of the carry out when you shift the dividend left. In fact, you
have to do more than that: you have to instantiate it into a value that
can be subtracted from.

To be able to calculate a full 32-bit divide (rather than the 31 bits of
the published fig-FORTH routines, eg) you have to, when checking for
underflow from a subtraction of the divisor from the dividend, do a
17-bit subtract - which takes three subtracts on an 8-bit machine.

This is a subtle fact, but it makes the algorithm actually work. ;-)