Added very basic USB support for Darwin/OSX
This commit defines the same words (for Darwin/OSX, via the IOKit) and libusb
defines for Linux/BSD.

The configuration is a bit lame, and I may improve it. I'm also considering
splitting it out into separate files. But before futzing more with it I wanted
to check it in.

It works - I was able to find and connect to a JS16, and do a CRC check. (It
failed - as it was supposed to. ;-)