Got 48M PLL clock working on JS (MCG)
This is the first step towards a USB chat loader - the xtal clock
working properly.

This first step is simply to run the serial chat "byteloader" at 24M
(bus clock) from the PLL (at 48M - which is how it will be set up for
USB) to test the code that initialises the clock.