Working USB enumeration code!! Woot!!
I spent a _lot_ of time the past two days reading the USB spec (for the
fourth time, roughly) and then poring over the disassembled JS16 boot
rom, and filling in some of the blanks.

I based some of code - structurally - on theirs, and it seemed like a
good fit. I had only two bugs: I had left in a longer (32 bytes instead
of 16) declaration of bMaxPacketSize0 in the device descriptor (which I
had inserted for testing), and I got the sense wrong of how to clear an
important interrupt flag bit (TOKDNE). Some things you clear to clear,
and others you set to clear. So I screwed that one up (but a related bit
in the same register I got right. Go figure.)

My code enumerates on Linux and the Mac. I can also do basic control
transfers (like reading the descriptors) on the Mac, but when I tried on
Linux I got a libusb error. So we're not completely out of the woods,
but this is an AWESOME milestone.