Make it easier to flash JS16's with USB code
I had already flashed them with special 24M versions of the chat code, so
switching to a ram version of the normal chat code (as a preparation for
flashing the last page) wasn't possible without this change to add the
xtal setup (for the 24m PLL clock) to the chat code.

Once running from RAM I loaded the USB firmware and flashed that.