S08 flash programming now supported over USB!
I had to do some surgery on the memory allocation code, and break S08flash
into two files, but it works great, and is _very_ fast.

I was able to load a ram-resident version of the USB code - it only consists
of handlers for vendor requests, since the chip has already enumerated - and
running from that I re-flashed the boot sector.