Huge rework of "universal chat" primitives
Structuring the code - both on the host and target - for flashing S08
parts has been a challenge. I've tried it several ways, and none felt right.

When thinking about how to do it in the context of USB, a light bulb went on.
The issue before had always been, "where do I store the bytes that I want to
flash?" Well, USB solves that: they are sitting in the OUT buffer, waiting to
be copied somewhere. My insight was that instead of just copying each byte, I
could burn it to flash. I've already got a loop running; instead of simply

  0 ,x sta

call a routine that does

  0 ,x sta

followed by the magic incantation to write the flash. Since the only code that
needs to be in RAM is the above routine, I can easily switch between copying
and flashing, and the code overhead is nil. It's also blazing fast.

So I took that insight (flash while doing something else) and tried to apply
it to the chat and BDM code, and in the process I realised that my
abstractions were wrong. My exported abstraction was about streaming bytes in
or out of the target, but that doesn't fit well with USB.

My new abstractions are about moving chunks back and forth (read, write,
flash), which in the case of the serial chat code can be easily expressed with
the streaming protocol that the target gives me. And by making one tiny
change, writing to flash with the serial chat is as easy as saying
"c.FlashNext". And it runs at wire speed! (It takes 45us to flash a byte, and
83.3us to send a character at 115200 bps.)

Unfortunately, flashing via BDM got slower (again ;-) because we no longer
loop in the executed code, so there is a lot of overhead wrangling the memory
and registers over BDM. In fact, it's no slower to run the entire show over
BDM - executing no code on the target.

The speed of flashing via BDM, while frustrating, is only a tiny concern,
though: the issue is with serial port latency, not something inherent to BDM.
Over USB it will be much faster. And in any case, I'm just using BDM to
breathe first life into the parts: I burn a boot/chatloader (serial or USB)
and leave BDM behind unless I brick the part and need to recover it. So I can
tolerate a couple of seconds to write a bootloader every once in a while...