USB targets are now "first class" chat targets!
What this means:
 * register display works
 * memory dump works
 * calling out to code (from interactive prompt) works

The main piece that was missing (before this commit, that is) was proper
display of registers, and a good way to set them, then run some code,
then re-capture their values, and then display them. This has been
standard issue on serial and bdm chat targets... but I didn't see a good
way to do it over USB.

The solution was to add a vendor request to read the SP register, and
also to "wrap" the loop that calls process-usb with a loop that calls
emulated-swi (which, like its namesake, pushes the PC and all the
registers onto the stack). In this way the interactive session can show
and change the captured registers, and when it comes time to execute
code, by putting a new value into the PC slot (and any other useful
values, like a data stack pointer into HX, if calling into Forth);
these changes are made using the read-memory and write-memory vender
requests. To execute, the "run" request simply executes an rti
instruction, which pops all the registers, and starts execution.

By cleverly leaving a return address on the stack, _above_ the SWI/RTI
register frame, we make it easy for the executed code to "return home":
it simply executes rts, and it re-enters the emulated-swi loop, which
then calls process-usb in a loop, and we're back in (interactive)
business.

I was lucky that it all fit. All this added about 20 bytes of code,
bringing me dangerously close to the limit... I think I have about six
bytes to spare, not counting the eight or so that I could reclaim by
replacing a few uses of compare immed then beq with a single instruction
(but I haven't figured out how to add it to assembler yet...)

So, there you have it! USB has officially arrived!