Saved a few more bytes in the USB code
By compressing the code even further - moving code to set IN's to NAK into a
routine and calling it instead - I squeezed out a few more bytes.

Don't Repeat Yourself!

Coding in the Forth style is generally really good at extracting all the
redundancy from a program. In this case I was a bit lazy, and didn't see the
possibilities for sharing.