Rewrote chat byteloader in "forth style" - like USB byteloader
This is an experiment. The code works, but it's not the "default" and I
haven't flashed any parts with it.

The main difference between this code (chat v3) and the existing code
(v2) - there was a v1 but only the host-side code still exists - is that
it is written as a subroutine that gets called repeatedly, rather than
as a loop that once entered is only exited via rti.

Both approaches have their merits. The big loop means that nothing
prevents us from moving the stack pointer and pushing arbitrary stuff on
it: what's sitting on the stack is an SWI/RTI frame _not_ a return
address. So the host can change the SP, fill up the new frame with
register values, and cause an RTI, and it work fine. In fact, this is
precisely how I used to flash the code. I had to move the SP (by command
from the host) because I was pushing the bytes that I wanted to program
onto the stack.

This approach, using a subroutine, is both more and less flexible. Less,
because changing the SP is hard - you have to copy the return address
somewhere, fiddle with SP, then push it back. And since the chat code
assumes that HX is precious, it starts to get pretty involved.

OTOH, a nice tail-recursive styles becomes possible, and to return to
the dispatch point, any code just needs to execute rts, or jmp to code
that returns.

It also becomes possible to do something I have been thinking about for
a while - having the byteloader code dispatch _more_ than just the
commands it knows about (BDM over serial, eg). Right now the way I
"multiplex" BDM and chat commands is pretty ugly. Having the flexibility
to call several routines to figure out what to do with a received
command byte is useful.

That said, none of this "mainline" code right now. I've set it aside for
a rainy day project... ;-)