Added support for S08JS USB ROM bootloader
It's now possible to flash bare devices (as long as they are JS8 or
JS16) via USB, without recourse to serial chat or BDM or any other
nonsense that only I have access to. ;-)

All you have to do, to (for example) load my USB chat firmware into a
bare device (assuming it's wired up with a crystal and USB connector and
the minimal hardware required!) is the following:

 * jumper device for bootloader (pull pin 7 (BLMS) low)
 * run the following command to start muforth:

  ./muforth -f target/HC08/load-S08usb-chat.mu4

It will autoprobe for the JS device and load the bootloader if found.
You probably want to be in hex to see the status messages more usefully,
so do

  hex

and then type

  flash-image

and it should write the image and reset the device. Rejumper it (and
probably disconnect/connect, without this YMMV) and reload the same
muforth file, and it should now say something about USB chat loaded, so
type

  chat

and you should see a dump of the registers.

  0 du

will get you into the memory dumper; further <Return> or <BS> to go up
or down. q to quit.

You're talking to the device!