File loading can now durably set radix and mode
And the HC08 build tools default to hex and __meta.

I realised that my saving and restoring of several settings - radix and
interpret mode among - was super _useful_, but not universally applicable. Now
it's possible to choose between raw-load-file (which saves very little) and
load-file (which continues to save just about everything).

By defaulting -f (the command-line flag to load a file) to call _raw_ load,
and by having all the HC08 project load files call build.mu4 with the
_non-raw_ load, it's possible to have build.mu4 set up things in a durable way
so that we always end up in hex and __meta, which is mostly what you want.

You can also do this on the command line, after the -f <file> but it's tedious
and I always forget to do it. This is much more fool/drool-proof.