Another rework of HC08 project load files
Things are even simpler - again.

I kept the names of the 908-specific load files, but changed all the S08 ones
to "bare" names with no mention of "S08".

The build.mu4 file is much smarter about loading support for flash and usb now
too. To create a new project you now just create a file with

  ld! target/HC08/build.mu4  ( ld! let's build set a nice radix and mode)
  ld target/HC08/myproject.mu4

and load it!