Nicer USB device recognition
Now if a USB device is found that is known to be an "interactive target"
(ie, running my chat firmware) it is queried for its S08 device id
(which is built-in and can be read from the register at 1806/7 hex) and
the proper device support gets loaded.

I'm currently using one USB product id for all interactive targets:
15a2:c0de (I've temporarily "borrowed" Freescale's vendor id for testing
;-). I could assign a different product id to each chip.

That could help ameliorate a minor annoyance: the Freescale S08 device
ids are _shared_ among related members of a sub-family (like JS8 and
JS16, or JM8 and JM16), so it's impossible to know how much memory the
target has.

The important thing is that the right device "equates" - for the
locations of registers etc - are loaded, and they are.