Added commented out Get_status USB request
So, this is where my beginner's luck (wrt USB) fails me.

I thought I should add this request - it's in the JS16 bootloader ROM,
which fares fine both on Linux and the Mac.

But when I add this seemingly benign code to respond to the Get_status
request, very bad things happen. Linux enumerates the device
sucessfully, but shortly after gets a Protocol error, and the device
must be re-enumerated in order to communicate. The Mac crashes _hard_ -
when I plugged a device running this code into my Mac (running OSX 10.6)
- it died. The mouse stopped moving. Solidly dead.

Which is weird. OSX - based on Mach - should be much more robust to
failure than this. That's a big part of the promise of microkernel
architectures.

But worse is that I can see _no_ reason that this change should

(a) cause such bad/catastrophic behaviour, and
(b) not return the correct status value (lsusb -v does _not_ show the status as
0, even when the Get_status code is included in the firmware.)

Ah, the mystery that is USB.