Fixed problem with STALL handling
Previously, I would stall the endpoint, wait for the STALL handshake to
be sent, and then unstall the endpoint, but I neglected to ready the USB
SIE to receive another SETUP request... So the target got wedged.

Now I make sure to re-enable receipt of SETUPs.

This solves a deep mystery I had with adding a handler for the Get_status standard
request; after putting in the code, not only would Linux (lsusb -v) not
read the status correctly, but after running lsusb, Linux would complain
about a Protocol error.

And if I even so much as plugged such a device (with this "broken"
get_status support) into an OSX Mac, the Mac would crash hard: mouse no
longer moving, everything dead.

I haven't tried this change on the mac, but it works reliably on Linux,
and enumerates on Windows too.