Moved common flash security code to S08flash.mu4
I had copied a few things over for the JS bootloader support. Since I
don't want to maintain two copies putting somewhere common made sense.
Since splitting S08flash into two parts (target and host) it makes to
put this in the target side, since it's about the chip, and not how we
talk to it.