Added a threaded version of the kernel - unfinished & untested
As I was pondering does> and ;code in native-code land, I thought it
might be an "interesting exercise" to write a threaded version of the
kernel. This is quite a departure, since from almost day one of this
project I decided that threading would be too slow, and I did a _lot_ of
work figuring out how to make a native version that was reasonably small
and fast.

There is something appealing about threaded forths, though. So I started
down this road, and got discouraged, but I'm not done yet. I'm checking
in this work, and hope to return to it - maybe once the metacompiler is
finished and I'm running "real forth" on the native side.

I'd love to be able to do side-by-side comparisons of execution speed
and code size...