90% of support for compiling literals specially
Now several operations - + and or xor < u< (do) and (+loop) - are
compiled with _two_ entry points: one for the case when both operands
are on the stack, and one for when the top operand is in "literal" form:
the low half in the A register, and the high half in xh (on the stack).

Currently I don't handle subtract efficiently (it always uses stack
entry point) and u< and < - being relational ops - _should_ have calls
to make-flag compiled after their calls; this hasn't been added yet. I
need a forward reference - a variable - for the address of make-flag.