Meta-compiler more careful about looking back at compiled code
The few optimisations I'm trying to do mostly - mostly concerning
treating literals specially - are implemented by looking back at
previously compiled code, and either changing or uncompiling it,
depending on the situation.

The previous checkin was pretty sloppy about identifying code snippets,
and very sloppy about checking that what it was looking was really the
last thing compiled. (It used to only capture calls and push-literal.)

Now any use of  code{  in the metac - which begins compilinig inline
assembler code - also compiles a call which, when the word is later
executed, will capture the  here  value before running the inline
assembler code. In this way, every piece of code compiled will clobber
last-code's earlier value, so it will _always_ point to the last
snippet. This way we can reliably look back and check things like: was
the last snippet a call? a call to a particular address? a literal push?
etc.

Tail call conversion is coming soon... ;-)