Support for common (;code-style) code for pushing constants
It's a bit (or a lot) slower than inline, native literals, but it's
also, on average, smaller: a routine that pushes a constant can be
called in two bytes if it's nearby, three otherwise. I can load many
constants in three or four bytes (worst case is five) but it takes
another three (jsr push-lit) to push...

I suppose I could make "native" versions of a few constant words, using
my existing load-constant code, but I'm thinking about ;code (does>)
style sharing, and being able to load constants this way is useful for
that endeavour.