Decided against nested interpreter modes; cleaned up S08 meta
All the pieces except for control structures are now in place in the
HC08 meta-compiler!

I decided that the fancy "meta-colon" modes - which are for writing
meta-compiler defining words in a "nice" syntax (without \m escapes, eg)
- make more trouble than they save. In particular, the ambiguities about
what mode  -]  should return to and what type of literal  ]  should
compile only arise because of the fancy meta-colon modes.

Now I need to grapple with control structures and compiling
if/while/until tests reasonably efficiently.