Moved tail-recursive ^ back into .ifndef threaded block
I was trying to keep all the native (vs threaded) meta code together,
but in one of my "cleanups" I moved ^ among its cousins ( ; in
particular) rather than keeping it separate. If/when I write a threaded
version of everything, I'd like to have as much as possible separated
already.