Solved timeout (protocol) errors in USB chat firmware
When launching code (via rti) I was making the mistake of returning a
status IN _before_ running the code. The host would then make a SETUP
request to get the SP from the target. The SETUP transaction would be
accepted by the target, but with long-running target code - such as the
do-loop and for-loop tests I was doing - the host would time out on the
IN (to read the SP).

It turns out that it works much better to _delay_ sending the status IN
packet of the run command until the target code finishes executing. I've
currently set the Linux usb layer to time out after 4sec - which is
plenty of time. Even doing 64 Ki loops takes perhaps 100ms.

With this small change, I can no longer break the connection when
trying long loops.