Be a bit more careful about wrapping addresses
Now not only is the word  +a  available (it adds and then wraps the
result to 64Ki) but now _any_ address applied to the HC08 memory image
is first wrapped/truncated.

And, going the other way, taking a "host" address (that is supposedly
inside the image) and converting it to a target address also wraps.