Synchronise RAM contents before executing code on target
We now keep track of what's been copied (in zram and ram regions) from
host to target. Every time we try to execute code on the target we first
synchronise to make sure it has all the right bits. This makes it
possible to do incremental, interactive coding and testing. (Or it will
be possible once __meta and __target are merged. Right now after
defining a new word, you return to __meta instead of __target, and have
to awkwardly switch back again.)