Initial attempt at doing BDM over USB
It works, but it brings out some structural problems in the BDM code.

For one, it needs to be factored into _transport_ support (over USB, over
serial), and _target_ support (S08, ColdfireV1). I'd like to support both
targets.

Also, my previous device support makes it impossible to select the chip type
that you want to talk to over BDM. Since a USB chip is recognised, it is
assumed that code will be built against it instead. This needs to be improved.

And lastly, I'm having the same problem switching between BDM and chat mode
using the USB transport as I had over serial. I need a good clean solution to
this. Right now to do BDM over USB a spawn a separate USB event loop which
only responds to BDM requests; a special "exit" request is required to return
to normal USB "chat" processing. This is clunky and error-prone.

In the process of making this change I realised that it would be nice to
default JS and JM parts to using a 24M crystal, esp when deciding which BDM
bit routines to use. So I added to their device descriptions the automatic
definition of the symbol "24m" which is checked when deciding which clock
speed to compile for.