BDM over USB!! at least a first stab
There are issues. ;-)

But the transports (over-SCI and over-USB) have been teased apart from
the command sets - making it easy to, among other things, add
ColdFire V1 BDM commands - but now I'm having confusion and issues with
the nested command sets (chat & bdm), with specifying which chip is
being compiled for (now that things are auto-id'd via USB), etc.

Lots more to do, but this is a beginning.