Combined SCI chat and USB chat flash routines
Before this change I was using two RAM-resident routines for writing to
flash: a simple one, for the SCI firmware) and an all-too-clever one,
used for the USB firmware.

I wanted to combine these into one routine because it was annoying
having multiples versions in multiple places. Now there is only one, and
it's _always_ in the same place (0100 hex).

The main change is the the flash command, instead of being an immediate
value that the host "pokes" into the correct location in memory, lives
in a zero-page byte - actually, the _last_ zero page byte (0ff). This
way the initial copying of RAM doesn't clobber the beginning of the
routine, which, in the too-clever self-modifying implementation, was
used both for copying RAM and writing to Flash...