Use EPCTL0 as the definition of USB support
Instead of defining a separate symbol (say, USB) for devices that support USB,
just look for EPCTL0 in the registers. We now use this to either flag loading
of code as impossible (ie, the core USB support) or to choose which version
(eg, BDM over serial versus BDM over USB).

This undoes some previous changes, but it's better.