Renamed bdm files to better reflect their function
The number 2 no longer appears in any of the BDM-over-USB files -- there is no
v2 of that, only v1.

The sci files have been renamed to use the word "serial" rather than sci and
to specify v1 or v2 of the protocol.

Build files have been updated to work as well.