Made it cleaner to AVR prog (kludge)
Before in order to run the AVR programming code, we had to load a bunch of
HC08 dev tools that have nothing to do with it -- except that they were needed
to _start_ the AVR programming-over-USB firmware.

I've added a kludge (as a proof-of-concept) to the chat-usb-host code to make
it easy to set the PC and start the code. Now very little code is loaded and
it's possible to run the programming code by simply saying "prog". No need to
"chat" first.

This process caused me to re-examine my choice to use GetRegs and SetRegs
-- and keeping a locally cached copy of the HC08 regs -- instead of getting
and setting them separately. It brings up lots of issues of what code has
"local knowledge" of the target's stack frame, etc... and since both the
chat-usb and chat-serial code use an SWI-style stack frame, right now the
knowledge of the stack layout is in one place -- in interact.mu4. But in this
case I'm not interacting with an HC08 target; I just want to set the PC and get
the damn thing running.

I may re-evaluate the whole thing. At the very least, since we're caching the
register frame, we should do it properly: we need a "valid" bit and a "dirty"
bit. It starts out invalid, so the first time we print the regs it goes to
fetch them. Every time we ask the target to Run, we check if the frame is
dirty, and push it to the target first (and mark it invalid).

If I axe the getregs/setregs idea, I still have to have code common to
chat-usb and chat-serial (and the 908 serial chat too) that knows how to, eg,
get and set the PC -- so it has to know the frame layout. Argh.