Added a way to show factory and alternate opcode names
Thinking about the AVR and PIC "factory" opcode names, some of which are
rather terrible (almost all, in the case of PIC).

This code tests out a way to define alternate names in the assembler, and a
way for the disassembler to use the dictionary names (factory or alternate)
when printing opcodes, instead of compiling inline strings (which is what I've
done in the past).