Changed serial and usb chat, bdm, and avrprog code to use pad
Instead of each defining its own buffer and so on. The only place that used
b>  and  b<  was bdm-usb-host, so I moved that code there.

Also, I defined little- and big-endian versions of w@ and w! (lew@ lew! bew@
bew!) so they exist no matter what kind of target is loaded. Using these and
pad was enough to do chat and avrprog pretty cleanly.