Fixed flashing S08s over BDM
Who knew the S08 flashing code was so broken?

Not only did it totally not work over BDM anymore, but there had always been
a bug w.r.t flashing the last page. If it wasn't blank, trims were saved, it
was erased, and trims and vectors flashed. But if it *was* blank, nothing
happened! So the trims and vectors wouldn't get flashed. Lame.

I decided the whole blank checking thing was a waste of time anyway. It's slow
over BDM, but erasing a page is really fast! So, whenever we're about to
program the first row of a page, we erase it, no matter what. (We save and
restore trims and vectors on the last page, as above.)

So that's fixed.

I also wasn't verifying the trims and vectors, and fixing that exposed a tiny
address math bug in the verify code (and there were four versions of it!?!) so
that ffff would display ffff_ffff_ffff_ffff. That's fixed, and now verifying a
chip that has trims burned onto shows up as a mismatch, which I think is fine.

Oh, and I had broken the JS bootloader flashing code too, when I renamed
something in build.mu4. So *that's* fixed. Boy. What a mess!