HUGE improvements to AVR programming code!
Only being able to move one byte per control transfer (really I think it took
two) made talking to the chips painfully slow. I think it took about a minute
to read out the flash of a 168.

By adding the ability to do "bulk" reads and writes -- of flash and eeprom and
even the signature bytes! -- it's much faster now. I can read the chip in 2
secs (approx) and program it in 4 secs. Could be faster? Probably. Right now
it only accepts a single IN or OUT buffer (16 bytes); by making it a bit
smarter I could send a whole page (128 bytes on mega168) in one control
request. I'm sure that would be somewhat faster, at the expense of a bit more
complexity in the firmware.

This change actually cleaned up the host-side code, too. I added some crufty
programming and reading tests. This needs to be reworked, fleshed out, etc,
but it fundamentally *works*.

On to make similar changes to the BDM over USB core for the HC08. It's
painfully slow too.